finitions

+++ src/main/java/springboot/PersonneRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource
-public interface PersonneRepository extends CrudRepository<Personne, Long> {
+public interface PersonneRepository extends CrudRepository<Personne, Integer> {
 
 }
